[DB-39487, DB-39488, DB-39501] Fix undesired logging, Undesired excep… (#85)
* [DB-39487, DB-39488, DB-39501] Fix undesired logging, Undesired exceptions for liveliness checks, Augment excpetion content

PINO was added to the Node.js driver for logging purposes.  It assumed the PINO property called "enabled" would be respected all the time, but it turns out to be overidden if the PINO level propery is set.  To protect from this, any place the PINO level property is set, we need to only do it if the enabled property is true, which is based on the environment variable PINO_LOG_ENABLED.

3DMessage as seeing exceptions trying to close exceptions during liveliness checks, but we don't have an exact use case.,  One feature of the Connection Pool is to check if connections are still active based on an interval.  It is possible that the interval check could occur during the point the pool is in the process of shutting down.  This commit will change the behavior of throwing an exception during a liveliness check when the pool is shutting down, to now just ignore the failure to close cleanly.

The C++ exception informaton for SQLCode and SQLState is now going to be added to the C++ Exception message that is passed for the Node.js exception created on the other side of the NaN bridge.  The format will be JSON, so the new codes and the existing text are passed as properties that can be parsed in javascript.

* Adding .env which changes not depend on the file existing

* fix code typo

* change build back to use latest